Abstracted logic for `OCRBoxModel`s
